move exoplayer javascript to a single file
